Move AuthController to the core api and the auth config to the application.yml file
