+++ java/sample-cqrs-events/src/main/java/com/amido/stacks/ApplicationConfig.java
@@ -3,7 +3,6 @@ package com.amido.stacks;
 import com.auth0.spring.security.api.JwtWebSecurityConfigurer;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Arrays;
 import org.apache.commons.lang3.BooleanUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -16,6 +15,8 @@ import org.springframework.security.config.annotation.web.configuration.WebSecur
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
 
 /** ApplicationConfig - Configuration class for Auth0 application security. */
 @Configuration("MySecurityConfig")
@@ -31,7 +32,7 @@ public class ApplicationConfig extends WebSecurityConfigurerAdapter {
   @Value(value = "${auth.apiAudience}")
   private String apiAudience;
 
-  @Value(value = "${auth.issuer}")
+  @Value(value = "${auth.issuer.https}")
   private String issuer;
 
   @Value(value = "${auth.isEnabled}")
