+++ java/sample-cqrs-events/src/main/java/com/amido/stacks/menu/api/v1/CreateMenuController.java
@@ -1,8 +1,8 @@
 package com.amido.stacks.menu.api.v1;
 
 import com.amido.stacks.core.api.dto.ErrorResponse;
+import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.menu.api.v1.dto.request.CreateMenuRequest;
-import com.amido.stacks.menu.api.v1.dto.response.ResourceCreatedResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
