+++ java/sample-cqrs-events/src/main/java/com/amido/stacks/menu/api/v1/UpdateCategoryController.java
@@ -12,11 +12,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import java.util.UUID;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 /** @author ArathyKrishna */
 @RequestMapping("/v1/menu/{id}/category/{categoryId}")
