+++ java/sample-cqrs-events/src/main/resources/auth.properties
@@ -1,8 +1,8 @@
-#The issuer of the JWT Token. Typically, this is your Auth0 domain with an https://
-auth.issuer=https://amidostacks.eu.auth0.com/
-
-#The unique identifier for your API
-auth.apiAudience=https://amidostacks.eu.auth0.com/api/v2/
-
-#Config switch to enable (true) or disable(false) Auth0.Provide either true or false to enable or disable
-auth.isEnabled=false
+##The issuer of the JWT Token. Typically, this is your Auth0 domain with an https://
+#auth.issuer=https://amidostacks.eu.auth0.com/
+#
+##The unique identifier for your API
+#auth.apiAudience=https://amidostacks.eu.auth0.com/api/v2/
+#
+##Config switch to enable (true) or disable(false) Auth0.Provide either true or false to enable or disable
+#auth.isEnabled=false
