+++ java/sample-cqrs-events/src/test/java/com/amido/stacks/menu/api/v1/impl/CreateItemControllerImplTest.java
@@ -16,8 +16,8 @@ import static org.springframework.http.HttpStatus.CONFLICT;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 import com.amido.stacks.core.api.dto.ErrorResponse;
+import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.menu.api.v1.dto.request.CreateItemRequest;
-import com.amido.stacks.menu.api.v1.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.menu.domain.Category;
 import com.amido.stacks.menu.domain.Item;
 import com.amido.stacks.menu.domain.Menu;
